requirements.txt
numpy~=1.23.5
Pillow~=9.5.0
matplotlib~=3.7.2